+++ flytekit/core/data_persistence.py
@@ -21,19 +21,16 @@ import os
 import pathlib
 import tempfile
 import typing
-from base64 import b64encode
 from typing import Any, Dict, Union, cast
 from uuid import UUID
 
 import fsspec
-import requests
-from flyteidl.service.dataproxy_pb2 import CreateUploadLocationResponse
 from fsspec.utils import get_protocol
 
 from flytekit import configuration
 from flytekit.configuration import DataConfig
 from flytekit.core.utils import timeit
-from flytekit.exceptions.user import FlyteAssertion, FlyteValueException
+from flytekit.exceptions.user import FlyteAssertion
 from flytekit.interfaces.random import random
 from flytekit.loggers import logger
 
@@ -323,80 +320,6 @@ class FileAccessProvider(object):
         except Exception as ex:
             raise FlyteAssertion(
                 f"Failed to put data from {local_path} to {remote_path} (recursive={is_multipart}).\n\n"
-                f"Original exception: {str(ex)}"
-            ) from ex
-
-
-class RemoteFileAccessProvider(FileAccessProvider):
-    def __init__(
-        self,
-        local_sandbox_dir: Union[str, os.PathLike],
-        raw_output_prefix: str,
-        data_config: typing.Optional[DataConfig] = None,
-    ):
-        super().__init__(
-            local_sandbox_dir=local_sandbox_dir,
-            raw_output_prefix=raw_output_prefix,
-            data_config=data_config,
-        )
-        self._get_upload_signed_url_fn = None
-        self._s3_to_signed_url_map: Dict[str, str] = {}
-
-    def put(self, from_path: str, to_path: str, recursive: bool = False):
-        if recursive:
-            raise AssertionError("Recursive put is not supported for remote file access provider")
-        to_path = self._s3_to_signed_url_map[to_path]
-
-        from flytekit.tools.script_mode import hash_file
-
-        md5_bytes, _ = hash_file(pathlib.Path(from_path).resolve())
-        encoded_md5 = b64encode(md5_bytes)
-        with open(str(from_path), "+rb") as local_file:
-            content = local_file.read()
-            content_length = len(content)
-            rsp = requests.put(
-                to_path,
-                data=content,
-                headers={"Content-Length": str(content_length), "Content-MD5": encoded_md5},
-            )
-
-            if rsp.status_code != requests.codes["OK"]:
-                raise FlyteValueException(
-                    rsp.status_code,
-                    f"Request to send data {to_path} failed.",
-                )
-
-    def get_random_remote_path(self, file_path: typing.Optional[str] = None) -> str:
-        if file_path and not pathlib.Path(file_path).exists():
-            raise AssertionError(f"File {file_path} does not exist")
-
-        from flytekit.tools.script_mode import hash_file
-
-        p = pathlib.Path(cast(str, file_path))
-        md5_bytes, _ = hash_file(p.resolve())
-        if self._get_upload_signed_url_fn is None:
-            raise AssertionError("_get_upload_signed_url_fn should be set before calling get_random_remote_path")
-        res = self._get_upload_signed_url_fn(content_md5=md5_bytes, filename=p.name)
-        res = cast(CreateUploadLocationResponse, res)
-        self._s3_to_signed_url_map[res.native_url] = res.signed_url
-        return res.native_url
-
-    @timeit("Upload data to remote")
-    def put_data(self, local_path: Union[str, os.PathLike], remote_path: str, is_multipart: bool = False):
-        """
-        The implication here is that we're always going to put data to the remote location, so we .remote to ensure
-        we don't use the true local proxy if the remote path is a file://
-
-        :param local_path: Local path to the file
-        :param remote_path: upload location
-        :param is_multipart: Whether to upload Directory or not
-        """
-        try:
-            local_path = str(local_path)
-            self.put(cast(str, local_path), remote_path, recursive=is_multipart)
-        except Exception as ex:
-            raise FlyteAssertion(
-                f"Failed to put data from {local_path} to {self._s3_to_signed_url_map[remote_path]} (recursive={is_multipart}).\n\n"
                 f"Original exception: {str(ex)}"
             ) from ex
