+++ flytekit/remote/remote.py
@@ -22,21 +22,24 @@ from datetime import datetime, timedelta
 import requests
 from flyteidl.admin.signal_pb2 import Signal, SignalListRequest, SignalSetRequest
 from flyteidl.core import literals_pb2 as literals_pb2
+from flyteidl.service.dataproxy_pb2 import CreateUploadLocationResponse
 
 from flytekit.clients.friendly import SynchronousFlyteClient
 from flytekit.clients.helpers import iterate_node_executions, iterate_task_executions
-from flytekit.configuration import Config, FastSerializationSettings, ImageConfig, SerializationSettings
+from flytekit.configuration import Config, DataConfig, FastSerializationSettings, ImageConfig, SerializationSettings
 from flytekit.core import constants, utils
 from flytekit.core.base_task import PythonTask
 from flytekit.core.context_manager import FlyteContext, FlyteContextManager
-from flytekit.core.data_persistence import FileAccessProvider, RemoteFileAccessProvider
+from flytekit.core.data_persistence import FileAccessProvider
 from flytekit.core.launch_plan import LaunchPlan
 from flytekit.core.python_auto_container import PythonAutoContainerTask
 from flytekit.core.reference_entity import ReferenceSpec
 from flytekit.core.type_engine import LiteralsResolver, TypeEngine
+from flytekit.core.utils import timeit
 from flytekit.core.workflow import WorkflowBase
 from flytekit.exceptions import user as user_exceptions
 from flytekit.exceptions.user import (
+    FlyteAssertion,
     FlyteEntityAlreadyExistsException,
     FlyteEntityNotExistException,
     FlyteValueException,
@@ -1004,6 +1007,11 @@ class FlyteRemote(object):
                             remote_logger.debug(f"Could not guess type for {input_flyte_type_map[k].type}, skipping...")
                     variable = entity.interface.inputs.get(k)
                     hint = type_hints[k]
+                    self.file_access._get_upload_signed_url_fn = functools.partial(
+                        self.client.get_upload_signed_url,
+                        project=project or self.default_project,
+                        domain=domain or self.default_domain,
+                    )
                     lit = TypeEngine.to_literal(ctx, v, hint, variable.type)
                 literal_map[k] = lit
 
@@ -1903,3 +1911,77 @@ class FlyteRemote(object):
             return remote_wf
 
         return self.execute(remote_wf, inputs={}, project=project, domain=domain, execution_name=execution_name)
+
+
+class RemoteFileAccessProvider(FileAccessProvider):
+    def __init__(
+        self,
+        local_sandbox_dir: typing.Union[str, os.PathLike],
+        raw_output_prefix: str,
+        data_config: typing.Optional[DataConfig] = None,
+    ):
+        super().__init__(
+            local_sandbox_dir=local_sandbox_dir,
+            raw_output_prefix=raw_output_prefix,
+            data_config=data_config,
+        )
+        self._get_upload_signed_url_fn = None
+        self._s3_to_signed_url_map: typing.Dict[str, str] = {}
+
+    def put(self, from_path: str, to_path: str, recursive: bool = False):
+        if recursive:
+            raise AssertionError("Recursive put is not supported for remote file access provider")
+        to_path = self._s3_to_signed_url_map[to_path]
+
+        from flytekit.tools.script_mode import hash_file
+
+        md5_bytes, _ = hash_file(pathlib.Path(from_path).resolve())
+        encoded_md5 = b64encode(md5_bytes)
+        with open(str(from_path), "+rb") as local_file:
+            content = local_file.read()
+            content_length = len(content)
+            rsp = requests.put(
+                to_path,
+                data=content,
+                headers={"Content-Length": str(content_length), "Content-MD5": encoded_md5},
+            )
+
+            if rsp.status_code != requests.codes["OK"]:
+                raise FlyteValueException(
+                    rsp.status_code,
+                    f"Request to send data {to_path} failed.",
+                )
+
+    def get_random_remote_path(self, file_path: typing.Optional[str] = None) -> str:
+        if file_path and not pathlib.Path(file_path).exists():
+            raise AssertionError(f"File {file_path} does not exist")
+
+        from flytekit.tools.script_mode import hash_file
+
+        p = pathlib.Path(typing.cast(str, file_path))
+        md5_bytes, _ = hash_file(p.resolve())
+        if self._get_upload_signed_url_fn is None:
+            raise AssertionError("_get_upload_signed_url_fn should be set before calling get_random_remote_path")
+        res = self._get_upload_signed_url_fn(content_md5=md5_bytes, filename=p.name)
+        res = typing.cast(CreateUploadLocationResponse, res)
+        self._s3_to_signed_url_map[res.native_url] = res.signed_url
+        return res.native_url
+
+    @timeit("Upload data to remote")
+    def put_data(self, local_path: typing.Union[str, os.PathLike], remote_path: str, is_multipart: bool = False):
+        """
+        The implication here is that we're always going to put data to the remote location, so we .remote to ensure
+        we don't use the true local proxy if the remote path is a file://
+
+        :param local_path: Local path to the file
+        :param remote_path: upload location
+        :param is_multipart: Whether to upload Directory or not
+        """
+        try:
+            local_path = str(local_path)
+            self.put(typing.cast(str, local_path), remote_path, recursive=is_multipart)
+        except Exception as ex:
+            raise FlyteAssertion(
+                f"Failed to put data from {local_path} to {self._s3_to_signed_url_map[remote_path]} (recursive={is_multipart}).\n\n"
+                f"Original exception: {str(ex)}"
+            ) from ex
